+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## dbt-databricks 1.0.3 (Release TBD)
 
+### Fixes
+- Make internal macros use macro dispatch pattern ([#72](https://github.com/databricks/dbt-databricks/pull/72))
+
 ## dbt-databricks 1.0.2 (March 31, 2022)
 
 ### Features